Rename imfile_handler to _file_handler
It is only used as a man-in-the-middle to load images / apply updates /
write to disk and imported only once in the __init__. It can therefore
be a private module of imutils.
