+++ vimiv/imutils/_file_handler.py
